+++ app/src/main/java/com/example/bodymanagerapp/menu/Exercise/ExerciseData.kt
@@ -1,5 +1,5 @@
 package com.example.bodymanagerapp.menu.Exercise
 
 class ExerciseData(val date : Int, val name : String, val set : ArrayList<Int>, val num : ArrayList<Int>?,
-                   val weight : ArrayList<Int>?, val time : ArrayList<String>?) {
+                   val weight : ArrayList<Float>?, val time : ArrayList<String>?) {
 }
